refactor(packaging): dev.Dockerfile and switch to Ubuntu
Multi-stage Docker files will attempt to build all stages, unless
--target is passed, even if they are not required for the unamed or
default stage. This addresses the issue, and results in a cleaner
release Dockerfile, so is a good improvement. The only awkward aspect is
we cannot inherit a stage from another Dockerfile, so a script is
included to build the base images before the dev.Dockerfile can be used.