chore: establish dev dependencies and scaffold Dockerfile
* chore: establish dev dependencies and scaffold Dockerfile

The Dockerfile scope is currently limited to `cardano-rosetta-server`,
however will be further developed to bring in the other services and
process management. The docker-compose is suited for development

Closes #2

* chore: add build script, add security param to download command in README

* fix(packaging): fix build stage as source of compiled code

I had mixed up the source for the built code. Was untested since the previous commit was scaffolding only

* fix: fixed build script to properly generate js code in `/dist` directory

* fix: fixed Dockerfile to make _offline-mode_ work

* refactor(packaging): dev.Dockerfile and switch to Ubuntu

Multi-stage Docker files will attempt to build all stages, unless
--target is passed, even if they are not required for the unamed or
default stage. This addresses the issue, and results in a cleaner
release Dockerfile, so is a good improvement. The only awkward aspect is
we cannot inherit a stage from another Dockerfile, so a script is
included to build the base images before the dev.Dockerfile can be used.

Co-authored-by: Alan Verbner <alan.verbner@iohk.io>