feat: cardano-serialization-lib added to the repo as a tar
While we wait for multiassets support on cardano-serialization-lib and
for it to be pushed to the npm registry, we will import as a library
that's pushed to the repo. Building and importing the repo is out of the
scope as the preferred way is to impor the library from the registry
itself.