Add support form building a aarch64 image
Updates cabal used to 3.6.2.0 (since this version has aarch64 binaries).

Uses TARGETARCH (set by docker to arm64 or amd64) to download the correct versions of cabal and ghc.

Installs numa and llvm which seem to be needed for building the aarch64 version of cardano-node (numa is also added to the runtime image).